BRICK-119-refactor: 카카오 시작버튼 훅으로 분리
