+++ components/SignUpStartForm.tsx
@@ -3,16 +3,12 @@
 import Image from 'next/image'
 import { Button } from '@radix-ui/themes'
 import kakaoTalk from '@/public/images/svgs/kakaoTalk.svg'
-import useWindowWidth from '@/hooks/useWindowWidth'
 import Link from 'next/link'
+import useKakaoButtonText from '@/hooks/useKakaoButtonText'
 
 function SignUpStartForm() {
-  const windowWidth = useWindowWidth()
+  const kakaoButtonText = useKakaoButtonText()
 
-  const kakaoButtonText =
-    windowWidth && windowWidth <= 400
-      ? '카카오로 시작하기'
-      : '카카오로 3초만에 시작하기'
   return (
     <div className="relative flex w-full max-w-480pxr flex-col items-center rounded-[.625rem] bg-gray-01 px-39pxr py-50pxr pt-139pxr mb:px-19pxr">
       <div className="w-full text-center">
